test for boolean value to vm fixes delete function

+++ app/src/main/java/com/learningwithmanos/uniexercise/heroes/source/local/HeroLocalSource.kt
@@ -1,7 +1,13 @@
 package com.learningwithmanos.uniexercise.heroes.source.local
 
+import android.util.Log
+import com.learningwithmanos.uniexercise.AppPreferences
 import com.learningwithmanos.uniexercise.heroes.data.Hero
 import com.learningwithmanos.uniexercise.heroes.data.LHero
+import com.learningwithmanos.uniexercise.heroes.ui.ApiViewModel
+import kotlinx.coroutines.DelicateCoroutinesApi
+import kotlinx.coroutines.GlobalScope
+import kotlinx.coroutines.async
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.map
 import javax.inject.Inject
@@ -28,7 +34,9 @@ interface HeroLocalSource {
 
     fun isEmpty(): Flow<Boolean>
 
-    suspend fun delete()
+    suspend fun fill(apikey: String, privatekey: String)
+
+    suspend fun setApi(apikey: String, privatekey: String)
 }
 
 class HeroLocalSourceImpl @Inject constructor(private val marvelDao : MarvelDao): HeroLocalSource {
@@ -50,8 +58,22 @@ class HeroLocalSourceImpl @Inject constructor(private val marvelDao : MarvelDao)
         return marvelDao.isEmpty()
     }
 
-    override suspend fun delete() {
-        marvelDao.deleteAll()
+    override suspend fun fill(apikey: String, privatekey: String) {
+        if (apikey != "0cf69d45e2482a87f2a9af138efba603" || privatekey != "8aa649a8b299924f9428f6db08189950b7bfd728") {
+            AppPreferences.apikey = "0cf69d45e2482a87f2a9af138efba603"
+            AppPreferences.privatekey = "8aa649a8b299924f9428f6db08189950b7bfd728"
+            marvelDao.deleteAll()
+            Log.d("Dispacher RUN", "fill: local db deleted API: $apikey Private: $privatekey")
+        }
+    }
+
+    override suspend fun setApi(apikey: String, privatekey: String) {
+        if (!(AppPreferences.apikey.equals(apikey)) || !(AppPreferences.privatekey.equals(privatekey))) {
+            AppPreferences.apikey = apikey
+            AppPreferences.privatekey = privatekey
+            marvelDao.deleteAll()
+            Log.d("Dispacher RUN", "setApi: local db deleted API: $apikey Private: $privatekey")
+        }
     }
 
     fun LHero.mapToHero() = Hero (
